Refactor + naming conventions
