+++ src/components/QrCodeCompanion/FormCheckBox.tsx
