+++ src/components/QrCodeCompanion/FormNumberInput.tsx
