+++ src/components/QrCodeCompanion/FormTextInput.tsx
